% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{cluster_load}
\alias{cluster_load}
\title{Overall cluster load}
\usage{
cluster_load(credentials = NULL)
}
\arguments{
\item{credentials}{Your credentials}
}
\description{
Overall cluster load for all clusters that you have access to.
}
