% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{dide_home}
\alias{dide_home}
\title{Network path construction}
\usage{
dide_home(path, username)
}
\arguments{
\item{path}{Path, relative to your home direcory}

\item{username}{Your username}
}
\description{
Construct network path to files in your home directory.
}

