% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{didewin_config}
\alias{didewin_config}
\alias{didewin_config_global}
\title{Configuration}
\usage{
didewin_config(credentials = NULL, home = NULL, temp = NULL,
  cluster = NULL)

didewin_config_global(...)
}
\arguments{
\item{credentials}{Path to credentials file, or username}

\item{home}{Path to network home directory, on local system}

\item{temp}{Path to network temp directory, on local system}

\item{cluster}{Cluster to use}

\item{...}{arguments to \code{didewin_config}}
}

