% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{login}
\alias{login}
\alias{logout}
\title{Log on to the cluster}
\usage{
login(credentials)

logout()
}
\arguments{
\item{credentials}{Either a list with elements username, password,
or a path to a file containing lines \code{username=<username>}
and \code{password=<password>} or your username (in which case
you will be prompted graphically for your password).}
}
\description{
Getting credentials without storing anything anywhere is
difficult, especially without storing information.  We try here.
}

