% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_mapping}
\alias{path_mapping}
\title{Describe a path mapping}
\usage{
path_mapping(name, path_local, path_remote, drive_remote)
}
\arguments{
\item{name}{Name of this map.  Can be anything at all, and is used
for information purposes only.}

\item{path_local}{The point where the drive is attached locally.
On Windows this will be something like "Q:/", on Mac something
like "/Volumes/mountname", and on Linux it could be anything at
all, depending on what you used when you mounted it (or what is
written in \verb{/etc/fstab})}

\item{path_remote}{The network path for this drive.  It
will look something like \verb{\\\\\\\\fi--didef3.dide.ic.ac.uk\\\\tmp\\\\}.
Unfortunately backslashes are really hard to get right here and
you will need to use twice as many as you expect (so \emph{four}
backslashes at the beginning and then two for each separator).
If this makes you feel bad know that you are not alone:
https://xkcd.com/1638 -- alternatively you may use forward
slashes in place of backslashes (e.g. \verb{//fi--didef3.dide.ic.ac.uk/tmp})}

\item{drive_remote}{The place to mount the drive on the cluster.
We're probably going to mount things at Q: and T: already so
don't use those.  And things like C: are likely to be used.
Perhaps there are some guidelines for this somewhere?}
}
\description{
Describe a path mapping for use when setting up jobs on the cluster.
}
\author{
Rich FitzJohn
}
