% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue}
\alias{queue}
\title{Create a queue object}
\usage{
queue(context, config = didewin_config())
}
\arguments{
\item{context}{A context}

\item{config}{Optional dide configuration information.}
}
\description{
Create a queue object
}
\details{
Queues are R6 objects with many methods.  They need documenting still.
}

