% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{shownodes}
\alias{shownodes}
\title{Node status}
\usage{
shownodes(cluster = NULL)
}
\arguments{
\item{cluster}{Name of the cluster}
}
\description{
Get status of the nodes.  This is just a holding place until the
website can return something like JSON.
}

