% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{submit}
\alias{submit}
\title{Submit tasks to the cluster}
\usage{
submit(task, name = NULL, cluster = NULL)
}
\arguments{
\item{task}{Filenames, as \emph{network paths}.  The function
\code{\link{dide_home}} can construct paths to your home
directory easily.}

\item{name}{Optional name for the task (scalar)}

\item{cluster}{Cluster to use}
}
\description{
Submit tasks to the cluster
}

