% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{web_client}
\alias{web_client}
\title{DIDE cluster web client}
\description{
DIDE cluster web client

DIDE cluster web client
}
\details{
Client for the DIDE cluster web interface.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{web_client$new()}}
\item \href{#method-login}{\code{web_client$login()}}
\item \href{#method-logout}{\code{web_client$logout()}}
\item \href{#method-logged_in}{\code{web_client$logged_in()}}
\item \href{#method-submit}{\code{web_client$submit()}}
\item \href{#method-cancel}{\code{web_client$cancel()}}
\item \href{#method-log}{\code{web_client$log()}}
\item \href{#method-status}{\code{web_client$status()}}
\item \href{#method-load_node}{\code{web_client$load_node()}}
\item \href{#method-load_overall}{\code{web_client$load_overall()}}
\item \href{#method-headnodes}{\code{web_client$headnodes()}}
\item \href{#method-r_versions}{\code{web_client$r_versions()}}
\item \href{#method-api_client}{\code{web_client$api_client()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an API client for the DIDE cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$new(credentials, cluster_default = "fi--dideclusthn", login = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentials}}{Either your username, or a list with at least
the element \code{username} and possibly the name `password. If not
given, your password will be prompted for at login.}

\item{\code{cluster_default}}{The default cluster to use; this can be
overridden in any command.}

\item{\code{login}}{Logical, indicating if we should immediately login}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\if{latex}{\out{\hypertarget{method-login}{}}}
\subsection{Method \code{login()}}{
Log in to the cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$login(refresh = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{refresh}}{Logical, indicating if we should login even if
it looks like we are already (useful if login has expired)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logout"></a>}}
\if{latex}{\out{\hypertarget{method-logout}{}}}
\subsection{Method \code{logout()}}{
Log the client out
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$logout()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logged_in"></a>}}
\if{latex}{\out{\hypertarget{method-logged_in}{}}}
\subsection{Method \code{logged_in()}}{
Test whether the client is logged in, returning \code{TRUE}
or \code{FALSE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$logged_in()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
Submit a job to the cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$submit(
  path,
  name,
  template,
  cluster = NULL,
  resource_type = "Cores",
  resource_count = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{to the job to submit. This must be a windows (UNC) network
path, starting with two backslashes, and must be somewhere that
the cluster can see.}

\item{\code{template}}{The name of the template to use}

\item{\code{cluster}}{The cluster to submit to, defaulting to the value
given when creating the client.}

\item{\code{resource_type}}{The type of resource to request (either \code{Cores}
or \code{Nodes})}

\item{\code{resource_count}}{The number of resources to request}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-cancel}{}}}
\subsection{Method \code{cancel()}}{
Cancel a cluster task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$cancel(dide_id, cluster = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dide_id}}{The DIDE task id for the task}

\item{\code{cluster}}{The cluster that the task is running on, defaulting to
the value given when creating the client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
Get log from job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$log(dide_id, cluster = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dide_id}}{The DIDE task id for the task}

\item{\code{cluster}}{The cluster that the task is running on, defaulting to
the value given when creating the client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status"></a>}}
\if{latex}{\out{\hypertarget{method-status}{}}}
\subsection{Method \code{status()}}{
Return job status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$status(state = "*", cluster = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{The state the job is in. Can be one of \code{Running},
\code{Finished}, \code{Queued}, \code{Failed} or \code{Cancelled}. Or give \code{*}
for all states (this is the default).}

\item{\code{cluster}}{The cluster to query, defaulting to the value
given when creating the client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_node"></a>}}
\if{latex}{\out{\hypertarget{method-load_node}{}}}
\subsection{Method \code{load_node()}}{
Return an overall measure of cluster use, one
entry per node within a cluster.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$load_node(cluster = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cluster}}{The cluster to query, defaulting to
the value given when creating the client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_overall"></a>}}
\if{latex}{\out{\hypertarget{method-load_overall}{}}}
\subsection{Method \code{load_overall()}}{
Return an overal measure of cluster use, one
entry per cluster that you have access to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$load_overall()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-headnodes"></a>}}
\if{latex}{\out{\hypertarget{method-headnodes}{}}}
\subsection{Method \code{headnodes()}}{
Return a vector of known cluster headnodes. Typically
\code{\link[=valid_clusters]{valid_clusters()}} will be faster. This endpoint can
be used as a relativelyly fast "ping" to check that you are
logged in the client and server are talking properly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$headnodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r_versions"></a>}}
\if{latex}{\out{\hypertarget{method-r_versions}{}}}
\subsection{Method \code{r_versions()}}{
Return a vector of all available R versions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$r_versions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_client"></a>}}
\if{latex}{\out{\hypertarget{method-api_client}{}}}
\subsection{Method \code{api_client()}}{
Returns the low-level api client for debugging
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{web_client$api_client()}\if{html}{\out{</div>}}
}

}
}
