% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_joblog}
\alias{web_joblog}
\title{Get job log}
\usage{
web_joblog(config, dide_id)
}
\arguments{
\item{config}{A config object}

\item{dide_id}{A \emph{DIDE} id number}
}
\description{
Get the log of a command from the HPC server.  Note that logs are
only available for jobs that have finished or failed (not availab e
}

