% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_jobstatus}
\alias{web_jobstatus}
\title{Job status}
\usage{
web_jobstatus(x, cluster = valid_clusters()[[1]], state = "*", n = Inf)
}
\arguments{
\item{x}{Either a username or configuration object}

\item{cluster}{The cluster to check (default taken from
configuration if passed as \code{x}).}

\item{state}{State of jobs to check; default is all states ("*").
Other valid options are "Running", "Finished", "Queued",
"Failed" and "Cancelled".}

\item{n}{number of rows to return (default is Inf, which returns
all available jobs).}
}
\description{
Get job status from the cluster
}

