% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_login}
\alias{web_headnodes}
\alias{web_logged_in}
\alias{web_login}
\alias{web_logout}
\title{Log on to the cluster}
\usage{
web_login(config = didehpc_config())

web_headnodes()

web_logout()

web_logged_in()
}
\arguments{
\item{config}{A \code{\link{didehpc_config}} object.  If settings
have been mde global this can be omitted.}
}
\description{
Getting credentials without storing anything anywhere is
difficult, especially without storing information.  We try here.
}
\details{
The logic here will change as the general configuration approach
solidifies.  For now it's all a bit crap.
}

