% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_submit}
\alias{web_submit}
\title{Submit tasks to the cluster}
\usage{
web_submit(task, config)
}
\arguments{
\item{task}{Filenames, as \emph{network paths}.}

\item{config}{Configuration information, via \code{\link{didewin_config}}.}
}
\description{
Submit tasks to the cluster
}

