% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{cluster_load}
\alias{cluster_load}
\title{Overall cluster load}
\usage{
cluster_load(...)
}
\arguments{
\item{...}{Parameters to pass through to
\code{\link{didehpc_config}}, or a \code{didehpc_config} object
as the first (and only) argument.}
}
\description{
Overall cluster load for all clusters that you have access to.
}
