% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{didewin_config}
\alias{didewin_config}
\alias{didewin_config_global}
\title{Configuration}
\usage{
didewin_config(credentials = NULL, home = NULL, temp = NULL,
  cluster = NULL, build_server = NULL, shares = NULL, template = NULL)

didewin_config_global(...)
}
\arguments{
\item{credentials}{Either a list with elements username, password,
or a path to a file containing lines \code{username=<username>}
and \code{password=<password>} or your username (in which case
you will be prompted graphically for your password).}

\item{home}{Path to network home directory, on local system}

\item{temp}{Path to network temp directory, on local system}

\item{cluster}{Name of the cluster to use (one of
\code{\link{valid_clusters}()})}

\item{shares}{Optional additional share mappings.  Can either be a
single path mapping (as returned by \code{\link{path_mapping}}
or a list of such calls.}

\item{template}{A job template.  On fi--dideclusthn this can be
"GeneralNodes", "4Core" or "8Core", while on "fi--didemrchnb"
this can be "GeneralNodes", "8Core", "12Core", "12and16Core" or
"16Core".}

\item{...}{arguments to \code{didewin_config}}
}
\description{
Collects configuration information.  Unfortunately there's a
fairly complicated process of working out what goes where so
documentation coming later.
}

