% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_didehpc.R
\name{queue_didehpc}
\alias{queue_didehpc}
\title{Create a queue object}
\usage{
queue_didehpc(context, config = didehpc_config(), root = NULL,
  initialise = TRUE, sync = NULL)
}
\arguments{
\item{context}{A context}

\item{config}{Optional dide configuration information.}

\item{initialise}{initialise context and log in to the cluster on
queue creation (recommended on initial creation, but not
required if you want to check on jobs).}
}
\description{
Create a queue object
}
\details{
Queues are R6 objects with many methods.  They need documenting still.
}
