% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_didewin.R
\name{queue_didewin}
\alias{queue_didewin}
\title{Create a queue object}
\usage{
queue_didewin(context, config = didewin_config(), login = TRUE)
}
\arguments{
\item{context}{A context}

\item{config}{Optional dide configuration information.}

\item{login}{log in to the cluster on queue creation (recommended)}
}
\description{
Create a queue object
}
\details{
Queues are R6 objects with many methods.  They need documenting still.
}

