% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_didewin.R
\name{queue_didewin}
\alias{queue_didewin}
\title{Create a queue object}
\usage{
queue_didewin(context, config = didewin_config(), initialise = TRUE,
  rtools = NULL)
}
\arguments{
\item{context}{A context}

\item{config}{Optional dide configuration information.}

\item{initialise}{initialise context and log in to the cluster on
queue creation (recommended on initial creation, but not
required if you want to check on jobs).}

\item{rtools}{Indicates if Rtools is required (needed in cases
where you need a C/C++ compiler, such as using Rcpp's inline
support, or rstan).  The default will try to detect if it is
required.}
}
\description{
Create a queue object
}
\details{
Queues are R6 objects with many methods.  They need documenting still.
}

