% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_submit}
\alias{web_submit}
\title{Submit tasks to the cluster}
\usage{
web_submit(task, name = NULL, cluster = NULL)
}
\arguments{
\item{task}{Filenames, as \emph{network paths}.}

\item{name}{Optional name for the task (scalar)}

\item{cluster}{Cluster to use}
}
\description{
Submit tasks to the cluster
}

